feat/46

+++ Ssafit/src/main/java/com/ssafy/ssafit/model/service/CommentServiceImpl.java
@@ -13,14 +13,14 @@ import com.ssafy.ssafit.model.dto.Comment;
 import com.ssafy.ssafit.model.dto.CommentResponse;
 
 @Service
-public class CommentServiceImpl implements CommentService{
+public class CommentServiceImpl implements CommentService {
 
 	@Autowired
 	private CommentDao commentDao;
-	
+
 	@Autowired
 	private ReplyDao replyDao;
-	
+
 	@Transactional
 	@Override
 	public void create(Comment comment) {
@@ -39,30 +39,26 @@ public class CommentServiceImpl implements CommentService{
 		originComment.setContent(comment.getContent());
 		commentDao.updateComment(originComment);
 	}
-
-	@Transactional
-	@Override
-	public void delete(int id) {
-		commentDao.deleteComment(id);	
-	}
-
-	// TODO: post에 기능이 합쳐질 경우, 삭제 예정
-	@Transactional
-	@Override
-	public void deleteByPostId(int postId) {
-		replyDao.deleteReplyByPostId(postId);
-		commentDao.deleteCommentByPostId(postId);
-	}
-
+	
 	@Override
 	public List<CommentResponse> getListByPostId(int postId) {
 		List<CommentResponse> commentResponses = new ArrayList<CommentResponse>();
 		List<Comment> comments = commentDao.selectListByPostId(postId);
-		
-		for(int i = 0; i < comments.size(); i++) {
-			CommentResponse comment = CommentResponse.build(comments.get(i), replyDao.selectListByCommentId(comments.get(i).getId()));
+
+		for (int i = 0; i < comments.size(); i++) {
+			CommentResponse comment = CommentResponse.build(comments.get(i),
+					replyDao.selectListByCommentId(comments.get(i).getId()));
 			commentResponses.add(comment);
 		}
 		return commentResponses;
+	}
+
+	@Transactional
+	@Override
+	public void delete(int id) {
+		if(replyDao.selectCountByCommentId(id) != 0)
+			commentDao.updateDeletedComment(id);
+		else
+			commentDao.deleteComment(id);
 	}
 }
